refactor: extract LinkUIBlockInserter to separate file and consolidate BackButton
- Extract LinkUIBlockInserter component to block-inserter.js for consistency with page-creator.js
- Consolidate BackButton component into dialog-wrapper.js since it's only used there
- Fix import issues with privateApis in block-inserter.js
- Remove separate back-button.js file to reduce file count
- Clean up unused imports in link-ui.js after component extraction
- Fix JSDoc alignment in block-inserter.js
- Maintain all existing functionality while improving file organization

This completes the component extraction phase of the LinkUI duplication cleanup.
